% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean the raw NOAA earthquake data}
\usage{
eq_clean_data(df_raw)
}
\arguments{
\item{df_raw}{A dataframe of raw NOAA data}
}
\value{
Returns a cleaned dataframe
}
\description{
This function takes the raw NOAA data and cleans it as per the requirements from
Coursera: 1) Creates a date column by uniting the year, month, day and converting it
to the Date class. 2)  The LATITUDE and LONGITUDE columns are converted to a numeric class.
3) A function eq_location_clean() is applied that cleans the LOCATION_NAME column.
We have also cleared any years that are negative as these create issues with as.Date.
}
\examples{
\dontrun{
  df <- load_data() \%>\% dplyr::filter(YEAR >= 0) \%>\%
       dplyr::mutate(date = get_date(DAY, MONTH, YEAR))
  print(class(df))
  head(df)
}

}
