% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create an earthquake label column}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{The dataframe that we'll be using to create our annotation column.}
}
\value{
Returns a character vector of earthquake details that we can use for
map annotations.
}
\description{
This function takes a dataframe (df) and returns a character vector
of HTML column names. Note that the dataframe must have columns
CLEAN_LOCATION_NAME, EQ_PRIMARY and DEATHS.
}
\examples{
\dontrun{
map2 <-load_data() \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(date) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
print(map2)
}
}
