% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans the LOCATION_NAME column}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{a dataframe containing a column LOCATION_NAME that will be cleaned}
}
\value{
Returns a cleaned dataframe as described
}
\description{
As per the instructions on Coursera: write a function eq_location_clean() that
cleans the LOCATION_NAME column by stripping out the country name (including the colon) and
converts names to title case (as opposed to all caps). This will be needed later for
annotating visualizations.
}
\examples{
\dontrun{
  df <- load_data() \%>\% eq_location_clean()
  print(class(df))
  head(df)
}

}
