% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{eq_map}
\alias{eq_map}
\title{Plot a leaflet map with annotations}
\usage{
eq_map(df, annot_col)
}
\arguments{
\item{df}{The dataframe of data that we wish to plot}

\item{annot_col}{The column from our input dataframe that we'll use as an
annotation column}
}
\value{
Returns an interactive leaflet map that you can print to the viewer.
}
\description{
We take a subset of our earthquake data and plot this on a map
using leaflet. See the leaflet docs (?leaflet) for a few relevant examples.
The input dataframe requires columns including LONGITUDE, LATITUDE, EQ_PRIMARY.
The annotation column ('annot_col') is also a required argument (no defaults).
}
\examples{
\dontrun{
 map1 <-load_data() \%>\%
   eq_clean_data() \%>\%
   dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(date) >= 2000) \%>\%
   eq_map(annot_col = "date")
 print(map1)
}
}
