% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Defines Our GeomTimeline class}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\description{
As per the course materials:
The ggproto() function is used to construct a new class corresponding to your new geom.
The geom_* function is constructed as a regular function.
Build a geom for ggplot2 called geom_timeline() for plotting a time line of
earthquakes ranging from xmin to xmaxdates with a point for each earthquake.
Optional aesthetics include color, size, and alpha (for transparency).
The xaesthetic is a date and an optional y aesthetic is a factor indicating
some stratification in which case multiple time lines will be plotted for
each level of the factor (e.g. country).
This geom looks to chart a timeline of earthquakes for a given country / countries
with points (in the example) representing earthquake events, point size indicating
earthquake magnitude and colour representing number of deaths. x (the date) is a
required aesthetic whereas y (country) is optional.
}
\examples{
\dontrun{
 df <- df_earthquakes \%>\% filter(COUNTRY \%in\% c("CHINA", "USA"), YEAR > 2000)
 ggplot(df, aes(x = date, y = COUNTRY,
                color = as.numeric(TOTAL_DEATHS),
                size = as.numeric(EQ_PRIMARY),
                label = CLEAN_LOCATION_NAME)) +
   geom_timeline() +
   labs(size = "Richter scale value", color = "# deaths") +
   ggplot2::theme(panel.background = ggplot2::element_blank(),
         legend.position = "bottom",
         axis.title.y = ggplot2::element_blank()) +
   ggplot2::xlab("DATE")
}
}
\keyword{datasets}
