% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Defines Our GeomTimelineLabel class}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelineLabel
}
\description{
As per the course materials:
The ggproto() function is used to construct a new class corresponding to your new geom.
The geom_* function is constructed as a regular function.
Build a geom called geom_timeline_label() for adding annotations to the earthquake data.
This geom adds a vertical line to each data point with a text annotation
(e.g. the location of the earthquake) attached to each line.
There should be an option to subset to n_max number of earthquakes, where we take the
n_max largest (by magnitude) earthquakes. Aesthetics are x, which is the date of the
earthquake and label which takes the column name from which annotations will be obtained.
Note that you should run geom_timeline first such that we have our points to annotate.
}
\examples{
\dontrun{
ggplot(df, aes(x = date, y = COUNTRY,
               color = as.numeric(TOTAL_DEATHS),
               size = as.numeric(EQ_PRIMARY),
               label = CLEAN_LOCATION_NAME)) +
  geom_timeline() +
  labs(size = "Richter scale value", color = "# deaths") +
  ggplot2::theme(panel.background = ggplot2::element_blank(),
        legend.position = "bottom",
        axis.title.y = ggplot2::element_blank()) + ggplot2::xlab("DATE") +
  geom_timeline_label(data=df)
}
}
\keyword{datasets}
