% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{get_date}
\alias{get_date}
\title{Create a date from NOAA data given DAY, MONTH, YEAR}
\usage{
get_date(days, months, years)
}
\arguments{
\item{days}{Day of the year as integer vector}

\item{months}{Month of the year as integer vector}

\item{years}{An integer vector of years}
}
\value{
A vector of the date objects returned from passing date strings to as.Date
}
\description{
The NOAA data has columns called DAY, MONTH, YEAR. Part of the requirements for
cleaning this data is to create a date column. All of the years are provided, but some of
the months and days are NA. We'll replace NA days with 1st and NA months with Jan.
Also, note that as.Date does not play nice with negative years, so these have been removed.
Vectors days, months and years must be of the same length.
}
\examples{
\dontrun{
  df <- load_data() \%>\% dplyr::filter(YEAR >= 0) \%>\%
       dplyr::mutate(date = get_date(DAY, MONTH, YEAR))
  print(class(df))
  head(df)
}

}
