% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{load_data}
\alias{load_data}
\title{Load the raw NOAA data}
\usage{
load_data(path = file.path("data_raw", "signif.txt"))
}
\arguments{
\item{path}{A character string file path to your downloaded data source. Default is "data/signif.txt"}
}
\value{
Returns a dataframe result after calling readr's read_delim
}
\description{
This will load the raw tab delimited data previously downloaded from the NOAA website
which is currently sitting in the data folder.
}
\examples{
\dontrun{
  df <- load_data()
  print(class(df))
  head(df)
}

}
